[#21]Vue2+CompositionApi에 CompositionApi 관련 코드외에 삭제
