+++ vue2+compositionApi/src/components/Events.vue
@@ -5,30 +5,51 @@
 </template>
 
 <script>
-import { mapState } from 'vuex';
 import EventCard from '@/components/EventCard.vue';
+import useUserEvents from '@/composables/useUserEvents';
+import useEventTitleSearch from '@/composables/useEventTitleSearch';
+import {
+  toRefs,
+} from '@vue/composition-api';
 
 export default {
   components: {
     EventCard,
   },
+  props: {
+    perPage: {
+      type: Number,
+      required: true,
+    },
+    page: {
+      type: Number,
+      required: true,
+    },
+  },
+  setup(props) {
+    const { perPage, page } = toRefs(props);
+    const { events } = useUserEvents(perPage, page);
+    const { searchQuery, eventMatchingSearchQuery } = useEventTitleSearch(events);
+
+    return {
+      events: eventMatchingSearchQuery,
+      searchQuery,
+    };
+  },
+  data() {
+    return {
+      filters: {}, // 3
+    };
+  },
   computed: {
     filteredEvents() { // 3
       return [];
     },
-    eventsMatchingSearchQuery() { // 2
-      return [];
-    },
-    events() {
-      return this.event.events;
-    },
-    ...mapState(['event']),
   },
-  watch: {
-    // TODO 코드 구현 필요
-    perPage: 'getUserEvents', // 1
-    // TODO 코드 구현 필요
-    page: 'getUserEvents', // 1
+  methods: {
+    updateFilters() { // 3
+
+    },
   },
 };
 </script>
